split engines to multiple modules
